Separate and internalize marshalers, and test exported surface only.
